+++ packages/at_persistence_secondary_server/lib/src/notification/hive_at_notification_keystore.dart
@@ -193,6 +193,9 @@ class HiveAtNotificationKeystore
   }
 
   @override
+  Future<bool> exists(String key) async => isKeyExists(key);
+
+  @override
   int entriesCount() {
     return _getBox().keys.length;
   }